* Added EmitterInstanceRemoveListener to allow clean up particle system's Ogre::Node. * GpuParticleSystemWorld::stop has additional parameter 'bool removeEmitterWhenFinished' - particle emitters will be removed or just stopped emitting new particles (restart will activate stopped emitter, although all particles will be removed immediately). * Clean up.

+++ src/GpuParticles/GpuParticleSystemWorld.cpp
@@ -24,46 +24,15 @@
 #include <OgreCompositorNode.h>
 #include <OgreCompositorWorkspace.h>
 #include <OgreImage2.h>
-#include <OgreSceneManager.h>
 
 #include <OgreLogManager.h>
 
 #include <GpuParticles/GpuParticleAffectorCommon.h>
 
-//#include <Other/DebugVariable.h>
-
 #include <GpuParticles/Hlms/HlmsParticle.h>
 #include <GpuParticles/Hlms/HlmsParticleDatablock.h>
 
 using namespace Ogre;
-
-namespace
-{
-//    struct ParticleData
-//    {
-//        /// Position
-//        float px, py, pz;
-
-//        /// Rotate xz
-//        float r;
-
-//        /// Colour
-//        float cr, cg, cb, ca;
-
-//        /// Size
-//        float sx, sy;
-
-//        /// Sprite percentage: ceil is taken as result and is casted to uint
-//        /// (it is easier to interpolate that way).
-//        float spritePercent;
-
-//        float lifetime;
-//        float maxLifetime;
-
-//        float dirX, dirY, dirZ;
-//        float dirVelocity;
-//    };
-}
 
 /// Like thread group data, but with BucketSize.
 struct GpuParticleSystemWorld::BucketGroupData
@@ -93,7 +62,17 @@ struct GpuParticleSystemWorld::BucketGroupData
     Ogre::uint32 nextBucketParticleIndex = 0;
 };
 
-const int GpuParticleSystemWorld::ParticleDataStructSize = sizeof(float) * 17u; // count number of floats inside 'ParticleData' struct
+const int GpuParticleSystemWorld::ParticleDataStructSize =
+        sizeof(float) * 3u +  // Position x, y, z
+        sizeof(float) +       // Rotate sprite
+        sizeof(float) * 4u +  // Colour r, g, b, a
+        sizeof(float) * 2u +  // Size x, y;
+        sizeof(float) +       // Sprite percentage: ceil is taken as result and is casted to uint
+                              // (it is easier to interpolate that way).
+        sizeof(float) +       // lifetime;
+        sizeof(float) +       // maxLifetime;
+        sizeof(float) * 3u +  // direction x, y, z
+        sizeof(float);        // float dirVelocity;
 
 const int GpuParticleSystemWorld::EntryBucketDataStructSize = sizeof(Ogre::uint32) * 5u;
 
@@ -241,6 +220,9 @@ GpuParticleSystemWorld::~GpuParticleSystemWorld()
         delete mParticleRenderables[i];
     }
     mParticleRenderables.clear();
+
+    destroyAllEmitterInstances();
+
     destroyBuffers();
 
     for (size_t i = 0; i < mRegisteredAffectorList.size(); ++i) {
@@ -402,12 +384,12 @@ bool GpuParticleSystemWorld::canAdd(const GpuParticleSystem* particleSystem) con
     return true;
 }
 
-void GpuParticleSystemWorld::stop(uint64 instanceId, bool destroyAllParticles)
+void GpuParticleSystemWorld::stop(uint64 instanceId, bool removeEmitterWhenFinished, bool destroyAllParticlesImmediately)
 {
-    std::vector<int> indexesToRemove;
-
     std::pair<EmitterInstanceIdToListIndex::iterator, EmitterInstanceIdToListIndex::iterator> idItRange
             = mEmitterInstanceIdToListIndex.equal_range(instanceId);
+
+    std::vector<int> indexesToRemove;
     for(EmitterInstanceIdToListIndex::iterator it = idItRange.first; it != idItRange.second; ++it) {
         indexesToRemove.push_back(it->second);
     }
@@ -416,22 +398,48 @@ void GpuParticleSystemWorld::stop(uint64 instanceId, bool destroyAllParticles)
         size_t instanceListIndex = indexesToRemove[i];
         EmitterInstance& emitterInstance = mEmitterInstances[instanceListIndex];
         emitterInstance.mRun = false;
-        if(destroyAllParticles || emitterInstance.mParticleCount + emitterInstance.mParticleAddedThisFrameCount == 0) {
-            destroyEmitterInstance(instanceListIndex);
+
+        if(removeEmitterWhenFinished || emitterInstance.mRemoveWhenFinished) {
+            emitterInstance.mRemoveWhenFinished = true;
+            int particleCount = emitterInstance.mParticleCount + emitterInstance.mParticleAddedThisFrameCount;
+            if(destroyAllParticlesImmediately || particleCount == 0 || emitterInstance.mIgnore) {
+                destroyEmitterInstance(instanceListIndex);
+            }
+            else {
+                freeUnusedBuckets(emitterInstance);
+            }
         }
         else {
-            freeUnusedBuckets(emitterInstance);
+            if(destroyAllParticlesImmediately) {
+                emitterInstance.restartEmitter(false);
+            }
         }
+
     }
 }
 
-void GpuParticleSystemWorld::stopAll()
+void GpuParticleSystemWorld::stopAndRemoveAllImmediately()
 {
-    for (int i = mEmitterInstances.size()-1; i >= 0; --i) {
+    for (size_t i = 0; i < mEmitterInstances.size(); ++i) {
         freeBuckets(mEmitterInstances[i]);
     }
-    mEmitterInstances.clear();
-    mEmitterInstanceIdToListIndex.clear();
+
+    destroyAllEmitterInstances();
+}
+
+void GpuParticleSystemWorld::restart(uint64 instanceId)
+{
+    std::pair<EmitterInstanceIdToListIndex::iterator, EmitterInstanceIdToListIndex::iterator> idItRange
+            = mEmitterInstanceIdToListIndex.equal_range(instanceId);
+    for(EmitterInstanceIdToListIndex::iterator it = idItRange.first; it != idItRange.second; ++it) {
+
+        EmitterInstance& emitterInstance = mEmitterInstances[it->second];
+        if(emitterInstance.mRemoveWhenFinished) {
+            continue;
+        }
+
+        emitterInstance.restartEmitter(true);
+    }
 }
 
 void GpuParticleSystemWorld::init(uint32 maxParticles,
@@ -501,11 +509,17 @@ void GpuParticleSystemWorld::init(uint32 maxParticles,
 void GpuParticleSystemWorld::createEmitterInstance(const GpuParticleEmitter* gpuParticleEmitterCore,
                                                    const Ogre::Matrix4& matParent,
                                                    Ogre::Node* parentNode,
-                                                   Ogre::uint64 idCounter)
+                                                   Ogre::uint64 idCounter,
+                                                   bool burstAutoRemove,
+                                                   EmitterInstanceRemoveListener* emitterInstanceRemoveListener)
 {
     mEmitterInstances.push_back(EmitterInstance());
     EmitterInstance& emitter = mEmitterInstances[mEmitterInstances.size()-1];
     emitter.mGpuParticleEmitter = gpuParticleEmitterCore;
+    if(gpuParticleEmitterCore->mBurstMode && burstAutoRemove) {
+        emitter.mRemoveWhenFinished = true;
+    }
+    emitter.mEmitterInstanceRemoveListener = emitterInstanceRemoveListener;
 
     Ogre::Matrix4 matOffset;
     matOffset.makeTransform(gpuParticleEmitterCore->mPos, Ogre::Vector3::UNIT_SCALE, gpuParticleEmitterCore->mRot);
@@ -546,16 +560,53 @@ void GpuParticleSystemWorld::destroyEmitterInstance(int instanceIndex)
 
     EmitterInstanceIdToListIndex::iterator itToRemove = findEmitterInstanceIt(instanceIndex);
     EmitterInstanceIdToListIndex::iterator itToSwap = findEmitterInstanceIt(mEmitterInstances.size()-1);
+    int count = mEmitterInstanceIdToListIndex.count(mEmitterInstances[instanceIndex].mId);
+    bool isLastEmitterOfSystem = (count == 1);
 
+    mEmitterInstances[instanceIndex].onRemove(isLastEmitterOfSystem);
     mEmitterInstances[instanceIndex] = mEmitterInstances[mEmitterInstances.size()-1];
     mEmitterInstances.pop_back();
 
-    // First set value (if itToRemove == itToSwap then this order will work as well).
-    itToRemove->second = itToSwap->second;
-    mEmitterInstanceIdToListIndex.erase(itToSwap);
+    // Let itToSwap points to listIndex which was pointed by itToRemove.
+    // First set value, then erase itToRemove.
+    // If itToSwap == itToRemove nothing wrong happen.
+    itToSwap->second = itToRemove->second;
+    mEmitterInstanceIdToListIndex.erase(itToRemove);
 }
 
-Ogre::uint64 GpuParticleSystemWorld::start(const GpuParticleEmitter* emitterCore, Ogre::Node* parentNode, const Ogre::Vector3& parentPos, const Ogre::Quaternion& parentRot)
+void GpuParticleSystemWorld::destroyAllEmitterInstances()
+{
+    std::map<Ogre::uint64, int> counter;
+
+    for (size_t i = 0; i < mEmitterInstances.size(); ++i) {
+        Ogre::uint64 id = mEmitterInstances[i].mId;
+        std::map<Ogre::uint64, int>::iterator itFind = counter.find(id);
+        if(itFind == counter.end()) {
+            counter[id] = 1;
+        }
+        else {
+            ++(itFind->second);
+        }
+    }
+
+    for (size_t i = 0; i < mEmitterInstances.size(); ++i) {
+        Ogre::uint64 id = mEmitterInstances[i].mId;
+        int& systemEmitterInstancesCount = counter[id];
+        bool isLastEmitterInSystem = (systemEmitterInstancesCount == 1);
+        mEmitterInstances[i].onRemove(isLastEmitterInSystem);
+        --systemEmitterInstancesCount;
+    }
+
+    mEmitterInstances.clear();
+    mEmitterInstanceIdToListIndex.clear();
+}
+
+Ogre::uint64 GpuParticleSystemWorld::start(const GpuParticleEmitter* emitterCore,
+                                           Ogre::Node* parentNode,
+                                           const Ogre::Vector3& parentPos,
+                                           const Ogre::Quaternion& parentRot,
+                                           bool burstAutoDelete,
+                                           EmitterInstanceRemoveListener* emitterInstanceRemoveListener)
 {
     if(mRegisteredEmitterCoresSet.find(emitterCore) == mRegisteredEmitterCoresSet.end()) {
         OGRE_EXCEPT( Exception::ERR_INVALIDPARAMS,
@@ -581,12 +632,17 @@ Ogre::uint64 GpuParticleSystemWorld::start(const GpuParticleEmitter* emitterCore
     Ogre::Matrix4 matParent;
     matParent.makeTransform(parentPos, Ogre::Vector3::UNIT_SCALE, parentRot);
 
-    createEmitterInstance(emitterCore, matParent, parentNode, idCounter);
+    createEmitterInstance(emitterCore, matParent, parentNode, idCounter, burstAutoDelete, emitterInstanceRemoveListener);
 
     return idCounter;
 }
 
-uint64 GpuParticleSystemWorld::start(const std::vector<GpuParticleEmitter*>& emitters, Node* parentNode, const Ogre::Vector3& parentPos, const Ogre::Quaternion& parentRot)
+uint64 GpuParticleSystemWorld::start(const std::vector<GpuParticleEmitter*>& emitters,
+                                     Node* parentNode,
+                                     const Ogre::Vector3& parentPos,
+                                     const Ogre::Quaternion& parentRot,
+                                     bool burstAutoDelete,
+                                     EmitterInstanceRemoveListener* emitterInstanceRemoveListener)
 {
     Ogre::uint32 requiredBuckets = 0;
     for (size_t i = 0; i < emitters.size(); ++i) {
@@ -619,15 +675,20 @@ uint64 GpuParticleSystemWorld::start(const std::vector<GpuParticleEmitter*>& emi
 
     for (size_t i = 0; i < emitters.size(); ++i) {
         const GpuParticleEmitter* gpuParticleEmitterCore = emitters[i];
-        createEmitterInstance(gpuParticleEmitterCore, matParent, parentNode, idCounter);
+        createEmitterInstance(gpuParticleEmitterCore, matParent, parentNode, idCounter, burstAutoDelete, emitterInstanceRemoveListener);
     }
 
     return idCounter;
 }
 
-uint64 GpuParticleSystemWorld::start(const GpuParticleSystem* particleSystem, Node* parentNode, const Vector3& parentPos, const Quaternion& parentRot)
+uint64 GpuParticleSystemWorld::start(const GpuParticleSystem* particleSystem,
+                                     Node* parentNode,
+                                     const Vector3& parentPos,
+                                     const Quaternion& parentRot,
+                                     bool burstAutoDelete,
+                                     EmitterInstanceRemoveListener* emitterInstanceRemoveListener)
 {
-    return start(particleSystem->getEmitters(), parentNode, parentPos, parentRot);
+    return start(particleSystem->getEmitters(), parentNode, parentPos, parentRot, burstAutoDelete, emitterInstanceRemoveListener);
 }
 
 void GpuParticleSystemWorld::updateInstances(float elapsedTime)
@@ -648,14 +709,16 @@ void GpuParticleSystemWorld::updateInstances(float elapsedTime)
         if(!emitterInstance.mRun) {
             emitterInstance.mTimeSinceStopped += elapsedTime;
 
-            // check if emitter stopped long enough time
-            if(emitterInstance.mTimeSinceStopped >= emitterCore->getMaxParticleLifetime() || emitterInstance.mParticleCount == 0) {
-                destroyEmitterInstance(i);
-                --i;
-                continue;
-            }
+            if(emitterInstance.mRemoveWhenFinished) {
+                // check if emitter stopped long enough time
+                if(emitterInstance.mTimeSinceStopped >= emitterCore->getMaxParticleLifetime() || emitterInstance.mParticleCount == 0) {
+                    destroyEmitterInstance(i);
+                    --i;
+                    continue;
+                }
 
-            freeUnusedBuckets(mEmitterInstances[i]);
+                freeUnusedBuckets(mEmitterInstances[i]);
+            }
         }
 
         float maxParticleLifetime = emitterCore->getMaxParticleLifetime();
@@ -819,6 +882,47 @@ void GpuParticleSystemWorld::setIgnoreEmitterInstance(uint64 instanceId, bool ig
         EmitterInstance& emitterInstance = mEmitterInstances[emitterListIndex];
         emitterInstance.mIgnore = ignore;
     }
+}
+
+bool GpuParticleSystemWorld::getIgnoreEmitterInstance(uint64 instanceId) const
+{
+    // If one emitter is ignored, all from the same instanceId should be ignored as well.
+
+    EmitterInstanceIdToListIndex::const_iterator it = mEmitterInstanceIdToListIndex.find(instanceId);
+    if(it == mEmitterInstanceIdToListIndex.end()) {
+        // wrong situation.
+        return true;
+    }
+
+    return mEmitterInstances[it->second].mIgnore;
+}
+
+bool GpuParticleSystemWorld::isRunning(uint64 instanceId) const
+{
+    // Finished when all emitter instances with such id are finished.
+    bool running = false;
+    std::pair<EmitterInstanceIdToListIndex::const_iterator, EmitterInstanceIdToListIndex::const_iterator> idItRange
+            = mEmitterInstanceIdToListIndex.equal_range(instanceId);
+    for(EmitterInstanceIdToListIndex::const_iterator it = idItRange.first; it != idItRange.second; ++it) {
+        int emitterListIndex = it->second;
+        const EmitterInstance& emitterInstance = mEmitterInstances[emitterListIndex];
+        running = running || emitterInstance.mRun;
+    }
+    return running;
+}
+
+bool GpuParticleSystemWorld::isFinished(uint64 instanceId) const
+{
+    // Finished when all emitter instances with such id are finished.
+    bool finished = true;
+    std::pair<EmitterInstanceIdToListIndex::const_iterator, EmitterInstanceIdToListIndex::const_iterator> idItRange
+            = mEmitterInstanceIdToListIndex.equal_range(instanceId);
+    for(EmitterInstanceIdToListIndex::const_iterator it = idItRange.first; it != idItRange.second; ++it) {
+        int emitterListIndex = it->second;
+        const EmitterInstance& emitterInstance = mEmitterInstances[emitterListIndex];
+        finished = finished && emitterInstance.isFinished();
+    }
+    return finished;
 }
 
 Ogre::ReadOnlyBufferPacked* GpuParticleSystemWorld::getParticleBufferAsReadOnly() const
@@ -1517,7 +1621,8 @@ void GpuParticleSystemWorld::emitParticleCreateGpu()
             //            Ogre::uint32 firstParticle = (emitterInstance.mParticleArrayStart + emitterInstance.mParticleCount) % emitterInstance.mEmitterParticleMaxCount;
             //            Ogre::uint32 lastParticle = (emitterInstance.mParticleArrayStart + emitterInstance.mParticleCreatedCount + particlesToCreate - 1) % emitterInstance.mEmitterParticleMaxCount;
 
-            if(emitterInstance.isStatic() || mInitLocationInUpdate) {
+//            if(emitterInstance.isStatic() || mInitLocationInUpdate) {
+            if(mInitLocationInUpdate) {
 
                 if(!emitterInstance.mGpuParticleEmitter->mBurstMode) {
                     // Create whole bucket of particles ahead of time.
